Updating fluxcd version of nginx to 1.1.1
This change updates only the fluxcd version of nginx to 1.1.1.
The armada version remains with version 0.41.2 and is also supported.

Test Cases:

PASS: Verify that there are no changes to the armada rpm generated
PASS: Verify that the armada version of app uses nginx 0.41.2
PASS: Run the rpm build and verify that two packages are generated:
      stx-nginx-ingress-controller-helm-<version>.tis.noarch.rpm and
      stx-nginx-ingress-controller-helm-fluxcd-<version>.tis.noarch.rpm
PASS: Install the new package with kustomize and verify that nginx pods
      are deployed and running with success
PASS: Verify that the fluxcd version of the app uses nginx 1.1.1

Story: 2009836
Task: 44604

Change-Id: Icbabe97720eb7d0e8c8676ae2a18ec5afa62b053
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>