Add helm plugin for nginx 1.1.1 requirement
The new version of nginx helm chart has ipFamily value set to IPv4 as
default. When the cluster is IPv6 the helm chart fails to be installed.
This code is adding a system_override to correctly set the ipFamily
according to the system in order to ensure that it works in both IPv4
and IPv6 environments.

This is also correcting a bad helm chart reference in fluxcd
helmrelease.yaml file.

Test Cases:

PASS: Apply app with system application-apply and check it properly
      generates ipFamily override for that system (IPV4 or IPV6).
PASS: Apply app with system application-apply and check that it
      runs successfully.
PASS: ISO built and installed successfully. Checked that nginx is
      working.
PASS: Verify that the fluxcd app is also running.
PASS: Verify that the fluxcd app version is 1.1.1 and armada version
      is 0.41.2
PASS: Manually copied the helm overrides generated by the helm plugin
      to fluxcd system override and re-installed with success.
PASS: Verify that the app runs successfully in both IPv4 and IPv6
      systems.

Story: 2009138
Task: 44697

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I2187e5a1457d8417fd6bb6b61322fc2923413fd4